backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
redis==5.2.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
httpx==0.28.1
xmltodict==0.14.1
ofxparse==0.21
groq==0.15.0
python-dateutil==2.9.0
uuid7==0.1.0
